modules/person_grpc/requirements.txt
grpcio==1.40.0
grpcio-tools==1.40.0
protobuf==3.18.0
psycopg2-binary==2.9.1
six==1.16.0
SQLAlchemy==1.3.19
